faster imageClone

+++ ants/registration/resample_image.py
@@ -41,10 +41,9 @@ def resample_image(img, resample_params, use_voxels=False, interp_type=1):
     >>> finn = ants.resample_image(fi,(50,60),True,0)
     >>> filin = ants.resample_image(fi,(1.5,1.5),False,1)
     """
-
     if img.components == 1:
-        inimg = img.clone('float')
-        outimg = img.clone('float')
+        inimg = img.clone('double')
+        outimg = img.clone('double')
         rsampar = 'x'.join([str(rp) for rp in resample_params])
 
         args = [img.dimension, inimg, outimg, rsampar, int(use_voxels), interp_type]
